modified entities, added services and exceptions

+++ src/main/java/repositories/ProductRepository.java
@@ -12,7 +12,7 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
     List<Product> findByNameContaining(String name);
     List<Product> findByBrand(String brand);
     List<Product> findByBarCode(String name);
-    List<Product> findById(int id);
+    List<Product> findById(long id);
     boolean existsByBarCode(String barCode);
 
 }
